[#5022] Profile: subject_areas -> ArrayField
